+++ src/modules/metric/UDPMetricSender.h
@@ -10,7 +10,6 @@
 #include "core/Trace.h"
 
 #include <stdint.h>
-#include <SDL_platform.h>
 
 #if defined(_WIN32) || defined(__CYGWIN__)
 #ifndef NOMINMAX